pyproject.toml
[project]
name = "log-ai"
version = "0.1.0"
description = "MCP Server for extracting and querying system logs"
dependencies = [
    "mcp>=0.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "sentry-sdk>=2.0.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.0"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio"
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that require external services (SSH, syslog server, etc.)"
]
